Modal-Description Display Fix
- Fixed Modal not updating state when <Header>'s random button fetches new data
